Platform data zero eliminaton and tinkering.
Using only one half of the LFSR value could result in values of zero for
the platform -- not too good!  I could just substitute the default
value, but XORing the two halves seemed to give a better distribution of
platform values (at least according to the seat of my pants).  Of
course, both halves could be identical...so, I still have to test for
zero and substitute the default as necessary. :-)