Implement a basic scoring routine.
Scoring is based on a delay counter that is decremented once for
every frame.  When it reaches zero, the counter is reset and the least
significant digit of the current score is incremented.  If that digit
overflows, the counter ripples through the other digits in the score.
No attempt is made to account for a roll-over at this time.

Once the score is incremented, it is compared with the high score.
The comparison runs from the most significant digit to the least.
If the current score is found to be higher, then it is copied over to the
high score area.